Sandbox Menus and Scene Management
The objective of this push was to make significant progress in scene management, and to finish the UI functionality of the Sandbox gamemode. This push was mostly a success.
Originally, basic scene manager commands were going to be used, but after consideration and testing, the flow was interrupted by load times and scenes suddenly appearing/disappearing. I conducted some research on scene transitions, and found this video (https://www.youtube.com/watch?v=CE9VOZivb3I&t=14s) which teaches how to set up fading scenes in and out. Following this tutorial I implemented this into my own scene, and it worked well. However, a short problem encountered was that loading scenes would break lighting, and would not generate proper lightmaps on load. To combat this issue, I tried using additive scene loading with a frame of delay, hoping that the short delay in which the current scene unloads and the next scene loads will allow the new scene to use their directional light, however this did not work in the end. Browsing this forum (https://forum.unity.com/threads/loading-scene-additively-causes-change-in-lighting.511566/), I eventually found a solution that worked, in which auto generation was disabled. After this the scene transition worked flawlessly, and just needs to be integrated into the other scenes.
Concerning the sandbox menu functionality, spawning and deleting objects were completed easily. All that was done was creating a cube and sphere prefab and a material to suit the objects, and the objects are spawned on a button press. Objects are also deleted by identifying them with a tag. These functions work flawlessly, and all that's left is to add functionality to the map change button.
Overall this push was a success, with some more progress to be made to fully finish the functions.