Post Alpha Test 4 & 5 - Preparing for Beta Test with Scene Management
The objective of this push was to start with scene management, and to factor any feedback from the previous alpha tests, implementing them somehow. Overall, this push was a great success.
Focusing on scene management - most of the functions were very easy to implement using Unity's SceneManager Class, however, moving between scenes by itself was not polished. Additionally, while still with considerable delay, the loading times between scenes did not feel long enough to implement a loading screen. To remedy this issue, a quad asset with material that had an editable alpha colour property was used to fade in and out of scenes, loading a scene when the screen is completely black. An issue that arose with this however was ray interactors not rendering past the invisible quad object. Using another material for the ray would fix this, however this would lead to the rays being visible all the time. Instead, the quad was shrunken to a scale of 0, allowing the default lines to appear and disappear properly. The scene transition manager was then put on every scene, which worked flawlessly.
Some issues from previous alpha tests was that the target was too close, so the distance was adjusted accordingly. Some bugs also arose from the locomotion system, where the character controller hitbox would collide with the foil. This issue will be fixed in the near future.
Overall, this push was a success, and the beta test is ready to be rolled out.