Audio & Implementation of Feedback
This push aimed to add more components to the audio functionality of the program, and integrate some feedback from tests. The overall success of this push is to be determined with some tests on VR.
Implementing feedback in this push came in the form of integrating map changes. Importing an asset from the Unity Asset Store, colliders were fitted onto two separate halls that came in the asset bundle. After the colliders were fitted, logic to swap between them was implemented. Using similar logic to changing pages on some of the menus, changing the maps involved toggling one hall to active and the rest to inactive. An incrementor was used cycle between the three available options, and is reset to 0 once the incrementor value goes above 2. This feature works flawlessly and has been tied to the map change button on sandbox.
In regards to the audio component, the mixer and groups were implemented in a previous push. This push aimed to create the audio source components and logic to change volume and play the sound. On reflection, the slider value needs to properly reflect the volume even on start, and the scripts to play sword FX should be redone in a simpler form. Testing also needs to be completed. However, audio clips and all the components to properly set up audio has been implemented.
Overall, this push has been a mixed success, and will be tested on the VR headset in the near future.