More slight audio adjustments before VR Test
This push aims to do some small clean ups and prepare to test audio and this build on VR. This build was a success and will be tested.
In the previous push a comment was made about the overcomplicated scripts for playing audio for the foil. This has since been revised and is now one script on all bones of the foil. While not efficient, this solution is easy and works correctly. Components have been correctly configured now, and need to be tested on VR. No issues have come up when developing this push.