Last Minute Adjustments for Beta Test 2
This push aimed to do a once-over revision of the build before the Beta test later this day, attempting to find any bugs and add any necessary features before the build. This push was a success.
After some contemplation and testing the limits of the model in free play mode, height scaling was added. Using this guide (https://www.youtube.com/watch?v=WAVa_9ilIEQ) height scaling was added, and worked without flaw. Further testing is required, and will be conducted in the beta test.
In addition, a minor fix was made to the slider settings in home menu. A small test showed that while the slider would correctly show the expected value when changed, the slider would reset to a default position not reflective of it's values when the scene was reloaded. To accommodate this issue, a small revision was made to make the slider.value equal the playerpref values. This did not bring any obvious faults, and as such is marked as complete.
No other bugs were found.
Overall this push was a success, and has readied the beta test build.