Implementation of Beta Test Feedback
This push aims to implement some of the fixes recommended in the beta test 2 feedback. This push was a success and implemented an improved behind line detection for the targets gamemode, along with some more quality enhancements.
Some feedback gotten was the targets gamemode throwing an error for various reasons during a round when the menu values were changed. This was swiftly fixed by making the object inactive during rounds. This feature worked flawlessly, and has not caused any issues.
In regards to the new behind the line system, this uses a timer that resets and shows the message whenever the controller is out of bounds, but does not show it for the length of the timer like the old function does. This feature is a success, but some more testing is appropriate to ensure it works for all scenarios.
Additionally, a minor UI change to the scoreboard rounds the average time to 3 numbers and adds a title indicating the average react time. This should overall create a more transparent user experience.
Another issue was the locomotion system, allowing players to walk out of the map. This was resolved using a character controller and character controller driver, however two bugs were caused by this change. The first was the player moving backwards at random and rapid speeds consistently, which was later discovered to be an effect of having a previous collider in the model. Once the previous collider was deleted, the locomotion system worked properly. The second issue was frame drop spikes which happened in periodic intervals whenever the player moved. Using the profiler in unity, the main issue seemed to be the cpu not being able to load everything before a frame fast enough, and would instead stall the gpu which created a massive resource hog. Eventually, to check the whether the issue was caused by the editor, the program was built and the issue did not persist. While in the editor it may be slow, it will not perform like this in the final release.
Overall this push was a success with no issues, and the only things left are small touchups to the UI and some tutorials.