Tutorials Development and Bug Fixes
This push aimed to fix a few issues but mainly develop some user guidance on the levels. This push was a success.
The approach taken to develop tutorials was to use billboards with some information that allow users to quickly get accustomed to the game before playing. This involved finding commercially usable and modifiable images and pasting them onto canvases, and adding some functionality to scroll through canvas pages. Overall this works flawlessly.
Other than that, some minor bug fixes. This includes: a new binding for oculus headsets, mutually exclusive animation options and an exit application function. Earlier this week I acquired an oculus for testing, and realised there is no menu button. This led to creating a new binding, which will be tested in an acceptance test for sandbox mode.  Mutually exclusive animation options was some verbal feedback received, and that was easily solved using a Toggle Group Component, which solved the exact issue. The exit function was also easily solvable, with the Application.Quit(); method, though this doesn't work in the editor and is yet to be tested on the current build.
Overall, this push was a success, and almost signals the submission of the major work.