Acceptance Testing, XR Adjustments and Internal Documentation
This push aimed to finish acceptance testing, and start internal documentation. Progress is still underway, but the acceptance testing is complete and works well.
The test was conducted on a HTC Vive and Oculus Rift S. Since development was mostly on the vive, this worked flawlessly. When testing the Oculus, tracking was ok, but it did not read inputs other than the trigger. This was resolved by moving to the Unity XR plugin, which worked well.
Other than that, internal documentation in the form of developer comments. This will be finished at a later date.
Overall, the commit is a success.