File Cleaning and Minor Adjustments
This push aims to finalise everything in the program, involving deleting unnecessary files, redoing image assets and modifying some distance values.
From the acceptance tests, the oculus is only able to use the generic XR controller binds, so the grips have had to be used. This is functional, and is now reflective on the help menus.
File cleaning, and deleting unnecessary scenes, scripts and plugins have been conducted to try and maximise efficiency and minimise package size.
Also, some menus have been moved further back to accommodate for player comfort and distance.
Overall, this push is a success and marks the end of development before submission.